+++ main/src/com/tramchester/geo/GridPosition.java
@@ -4,22 +4,22 @@ package com.tramchester.geo;
 // https://en.wikipedia.org/wiki/Ordnance_Survey_National_Grid#All-numeric_grid_references
 
 public class GridPosition {
-    private final long eastings;
-    private final long northings;
+    private final int eastings;
+    private final int northings;
 
-    public GridPosition(long eastings, long northings) {
+    public GridPosition(int eastings, int northings) {
         this.eastings = eastings;
         this.northings = northings;
     }
 
     // National grid positions as always +ve
-    public static final GridPosition Invalid =  new GridPosition(Long.MIN_VALUE, Long.MIN_VALUE);
+    public static final GridPosition Invalid =  new GridPosition(Integer.MIN_VALUE, Integer.MIN_VALUE);
 
-    public long getEastings() {
+    public int getEastings() {
         return eastings;
     }
 
-    public long getNorthings() {
+    public int getNorthings() {
         return northings;
     }
